+++ jsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@PokemonFocus/*": ["./components/list-pokemon/pokemon-focus/*"]
     }
   }
 }
