+++ src/components/list-pokemon/Buttons.jsx
@@ -1,6 +1,11 @@
 import React, { useContext } from "react";
 import { PokemonsContext } from "@/context/PokemonsProvider";
 import "@/styles/buttons.css";
+/**
+ * Renders the Previous and Next buttons for navigating through the list of pokemons.
+ *
+ * @returns {JSX.Element} - The rendered component.
+ */
 const Buttons = () => {
   const { goNext, goPrev } = useContext(PokemonsContext);
   return (
