+++ src/components/list-pokemon/pokemon-focus/PokemonNotLoaded.jsx
@@ -1,0 +1,1 @@
+export const PokemonNotLoaded = () => <div>Not loaded Detail</div>;
