+++ src/pages/Homepage.jsx
@@ -2,7 +2,7 @@ import MemoProvider from "@/context/MemoProvider";
 import Container from "@/components/Container";
 import LayoutMemo from "@/components/memotest/LayoutMemo";
 import List from "@/components/list-pokemon/List";
-import PokemonFocus from "@/components/list-pokemon/PokemonFocus";
+import PokemonFocus from "@/components/list-pokemon/pokemon-focus/PokemonFocus";
 
 import "@/App.css";
 import TimerProvider from "@/context/TimerProvider.jsx";
